Temporary change to KeyboardInterrput handling
CTRL-C and CTRL-V are considered to be keyboard interrupts on Windows, which means that the user cannot CTRL-V a message into the program. The sys.exit() call for the KeyboardInterrupt error catcher is commented out. Possible solution is to define explicitly an exit function with a specific command input. Think about using flags?